requirements.txt
# Core Django
Django==5.0.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.2
python-dotenv==1.0.1
drf-yasg==1.21.7

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Static Files & Media
whitenoise==6.6.0
django-storages==1.14.2
boto3==1.34.69

# WebSockets
channels==4.0.0
channels-redis==4.2.0

# Celery for background tasks
celery==5.3.6
redis==5.0.1

# AI and ML - FIXED: use available version
openai==1.16.2
groq==0.9.0
ortools>=9.15.6755

# PDF Generation
weasyprint==62.3

# Monitoring
sentry-sdk==1.45.0

# Database Locking
django-pglock==1.0.0

# Utilities
openpyxl==3.1.2
Pillow==10.2.0
django-import-export==4.0.0

# Production Server
gunicorn==21.2.0